backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1